ui: Add appbar in AddNewRouteFragment
