--- app/src/main/java/com/maru/todayroute/addnewroute/AddNewRoute.kt
@@ -1,27 +1,0 @@
-package com.maru.todayroute.addnewroute
-
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.fragment.app.Fragment
-import com.maru.todayroute.databinding.FragmentAddNewRouteBinding
-
-
-class AddNewRoute : Fragment() {
-    private var _binding: FragmentAddNewRouteBinding? = null
-    private val binding get() = _binding!!
-
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View {
-        _binding = FragmentAddNewRouteBinding.inflate(inflater, container, false)
-        return binding.root
-    }
-
-    override fun onDestroyView() {
-        super.onDestroyView()
-        _binding = null
-    }
-}
