+++ app/src/main/res/menu/menu_add_new_route_toolbar.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto">
+    <item
+        android:id="@+id/menu_add_new_route_toolbar_save"
+        android:title="save"
+        android:icon="@drawable/ic_baseline_check_24"
+        app:showAsAction="always"/>
+</menu>
